+++ Boletin12/src/Tienda/Modelos/IArticulo.java
@@ -1,9 +1,9 @@
 package Tienda.Modelos;
 
 public interface IArticulo {
-    public boolean setPrecio(double precio);
+    public boolean setPrecio(float precio);
     public boolean setNombre(String nombre);
-    public double getPrecio();
+    public float getPrecio();
     public String getNombre();
     public int getId();
 }
